retiring only Vocab and preserving Vectors
