+++ torchtext/__init__.py
@@ -2,6 +2,7 @@ from . import data
 from . import nn
 from . import datasets
 from . import utils
+from . import vocab
 from . import experimental
 from . import legacy
 
@@ -15,6 +16,7 @@ __all__ = ['data',
            'nn',
            'datasets',
            'utils',
+           'vocab',
            'experimental',
            'legacy']
